requirements.txt
# Core ML and Data Science Libraries
pandas>=1.3.0
scikit-learn>=1.0.0
numpy>=1.21.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# DVC for ML Pipeline Management
dvc>=2.0.0

# Model Serialization
joblib>=1.1.0

# YAML Configuration
PyYAML>=6.0

# Jupyter Notebook Support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Development and Testing (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0